fix: renamed to universal and added autodetection of multiplexed logs
